Remove trailing slash from path in PATH.basename (#11938)
When calling PATH.basename on paths with a trailing slash (i.e. /emscripten/),
an empty string will be returned instead of "emscripten". This leads to errors
with rmdir and rename called on paths with a trailing slash.

This PR looks to first normalize the path, leading to at most one trailing slash
remaining in the path, and then removing that trailing slash.

Fixes #11897