Require python3 in test code
We alreedy require python3 everywhere but we are holding off breaking
python2 compat in emscripten for a few releases.

I don't see why we should do that for the test code itself so this
change makes the test code start to depend on python3.

And to show that I'm serious I thrown in a few f-strings.