Stop running --inline-main (#11995)
LLVM adding __original_main can be confusing for users, as it shows up in
stack traces, and they don't expect it. So we added a pass to inline that into
main, and we ran it on all builds, debug and release. However, this has some
downsides:

* As a comment that this PR removes shows, we actually stopped doing
  this in debug builds because it can mess with DWARF. So the place we need
  it most is no longer relevant anyhow, and it wasn't actually helping us with
  debugging.
* It is more work and complexity during link, which we are trying to reduce,
  see WebAssembly/binaryen#3043
* There are some planned changes to these functions and _start and all
  that anyhow (see discussion on the PR) so this will become irrelevant soon.

This does not affect behavior, unless you look at the wasm internals. But a noticeable
effect is that if you use ASYNCIFY_ADD or ASYNCIFY_ONLY then you may need
to add __original_main to there (since you are doing manual fine-tuning of
the list of functions, which depends on the wasm's internals) - see the test change
here.

Note that this should not matter in -O2+ anyhow as normal inlining generally removes
__original_main.