Fix __sys_fallocate with WASM_BIGINT (#12004)
I ran the full test suite with WASM_BIGINT to look for any issues. It
found just this one. The code assumed that the i64 was an actual i64,
but this uses the musl linux syscall convention, where the i64 is always
in two i32 values.

Add testing for that, and also a comment on another test that
fails with WASM_BIGINT but in an expected way.