Only run --no-exit-runtime when optimizing (#11998)
That pass is a pure optimization: it removes calls to atexit when they
would be ignored anyhow (EXIT_RUNTIME == 0).

This removes a warning in atexit's implementation that was never
actually called before: we used to always run that pass, so if
EXIT_RUNTIME == 0 then we never had any calls to atexit
anyhow. Now that we only run the pass when optimizing, leaving
that warning would be a noticeable change (and it broke some
tests actually!) so just remove it. With that, this is essentially
NFC except that non-optimized builds may be a little larger
(containing calls to atexit that end up doing nothing).

Helps WebAssembly/binaryen#3043