Remove unnecessary restrictions around RTTI from wire.h (#10914)
Fixes #6551

Removes the static_asserts that check if the type for a LightTypeID is_polymorphic.
AFAICT, LightTypeID works fine with polymorphic types, and provides the same
identity guarantees as when RTTI is enabled. Adds a test asserting such code
compiles and runs correctly.

I'm still not clear why this restriction was in place - it may have been necessary
in previous impelmentations, or perhaps with previous versions of LLVM/Clang.
However, as far as I can tell, it's not necessary now and only limits options
around doing non-RTTI builds with emscripten. Attempting to create a testcase
showing why the restrictions makes sense fails (see discussion in PR).