tests: always build dirdctly from input file where possible.
The `do_run` method takes a string and then writes it to temporary
file.   For tests that have source file inputs on disk, we want to
avoid reading this in and then writing it back out again because
in doing so the error messages when building test code becomes way
less useful.