Use direct table access over dynCall in invoke_xx functions (#11979)
This change starts the transition away from dynCall and towards
direct use of the wasm table.

In particular this change replaces the use of dynCall in invoke_xx
functions where possible.  dynCall is still needed for functions with
i64 in thier signature when WASM_BIGINT is not enabled.  When
WASM_BIGING is enabled this change removes all use of dynCall by
invoke_xx functions.