Make file packager --no-heap-copy always apply, and remove the option (#12027)
The old default was to copy a file package into main memory (into space
malloc'd for it). That was not memory-efficient as it meant the main memory
needed to be bigger, and also the memory could never be freed. While
mmap on that could be fast, we added the --no-heap-copy option to
keep file data on the JS heap and not copy it in. That uses less memory
but still is just as fast at fread etc.

This makes that newer option the default, and removes the option to do it
the old way. This is better for memory usage, and in any case people using
file data in a heavy way may be rolling their own solutions anyhow.

This also removes the file system usage of getMemory(), which would
perform allocation during startup from JS, which is a source of complexity
I am working to remove for WebAssembly/binaryen#3043