Refactor and simplify asyncify logic for calling back in (#12081)
This removes callStackIdToFunc, the map of ids to functions. Instead it
makes us always use the map of ids to function names. That removes one
of the two mechanisms and makes the code simpler, which will help in a
followup PR that adds support for direct table calls, which we need for
#12059.

I think we kept both mechanisms because I was worried about performance,
but I think I was wrong. After debugging for #12059 I think I have a better
understanding of things and it seems fine to use names - it's just a single
JS map lookup for asyncify operation, which is small compared to the
actual asyncify overhead.

Move some code to maybeStopUnwind, with no changes. That code
will be useful in the followup as well.

Also clean up some debugging code to make it more useful.