Use llvm-objcopy to strip DWARF debug info when not running wasm-opt (#12077)
Relative to / followup to #12075. This does the same for debug info, it gets
us back to the same output as before (even stripping it in unoptimized
builds if -g is not present) by using llvm-objcopy when wasm-opt isn't
being run anyhow.

Together these PRs make our work on WebAssembly/binaryen#3043
have less noticeable effect that could surprise users.

Note that wasm-ld has --strip-debug which we could use, but it will
strip both the Names section and DWARF sections, and we may want
just the Names section, so we can't just use that.