Document WebGL1 / WebGL2 context creation issues (#11839)
Even if both WebGL1 and 2 are supported in a build, we don't try both. If
the user requested 1, we try 1, and if 2, we try 2. The user can try to
create 2 and then fall back to 1 if desired.

The case where both 1 and 2 are supported in the build but the user
does *not* specify the version is a little ambiguous. We currently also
do not do any fallback there, but we try for 2 if it looks like it will
succeed (if WebGL2Context is defined - but that may not necessarily
work when actually creating the context).

Helps #11819