Enable and improve test_stack_overflow_check (#12095)
Depends on WebAssembly/binaryen#3091

This both enables the check, and adds a mode that checks many small
allocations instead of one huge one. That ensures we check both the lower
and upper limits of the stack region, and so are safe from the overflow
issue mentioned in #9039 (comment)

Also fix when the initialization happened - it was in callMain but that
didn't help cases without main, or using minimal runtime.

Also fix the skipping in test_core.py which looked for the old option.
With that fixed, wasm2ss (the test mode that runs with full stack
checks) all passes.

Fixes #9039 and re-enables `test_safe_stack_dylink` after the
binaryen roll + these fixes.