Switch to llvm's PIC ABI for relocatable code
This change replaces the current PIC ABI (where function and data
addresses are resolved via `fp$` and `g$` function calls with the PIC
ABI use by llvm which involved importing `GOT.func` and `GOT.mem`
mutable wasm globals.

This allows us to pass `--new-pic-abi` to wasm-emscripten-finalize
which disabled the PIC ABI transformation normally done by binaryen.

As part of this change we now store the `__sig` of each JavaScript
function on the function so it is available at runtime.  This
runtime signature is needed when a dynamically linked module
imports a JavaScript function by address.  In this case the function
must be added to the table, and for that to succeed we need to know
its signature.