Remove the use of the native JS optimizer: part 1
This change simply asserts that we native optimizer is not
being used for anything anymore.

As far as I can't tell use_native only has a single call site
and always returns False since its never passed `asm` as one
of the passes anymore.

Once this lands we can actually remove it.