Stop running wasm-opt at all in some -O0 builds (#12108)
This reaches the first milestone of WebAssembly/binaryen#3043, to stop running wasm-opt
in some -O0 builds (the next milestone is wasm-emscripten-finalize).

To do so, this stops running --post-emscripten unless we optimize. This was the last pass
we ran in -O0 builds, so it finally gets us into the code path of not running wasm-opt. That
turned up a bug in the logic there - view that area without whitespace, it's a trivial change
(we checked is passes is empty an extra time in an outer unnecessary scope).

By running wasm-opt less this appears to have uncovered an LLVM bug,

https://bugs.llvm.org/show_bug.cgi?id=47413

so part of one test is disabled.