Skip data-upload calls if size is zero
This implements the OpenGL ES spec behavior that these calls may be
valid (but no-op) if the upload size is zero.

Note this is still not quite conformant, because it will also skip
validation (which normally occurs in the browser) if the size is zero.
However, that problem is lower severity than this one.

Fixes #9370.