Pthreads: Fix proxying confusion between main browser and main runtime thread (#12102)
The main browser thread is the page with the DOM and HTML, while the
main runtime thread is where we start up the JS, and where things like
stdio proxying go (so all file access is on the same thread). Normally that
is the same place. However, a pthreads application started in a worker has
no access to the main browser thread. In that case emscripten_is_main_browser_thread
always returns false, while emscripten_is_main_runtime_thread returns
true (on the main thread).

This PR fixes some confusion there, some proxying code checked the
main browser thread when it meant to check the main runtime thread.
In that case we actually care about handling proxied calls. Without this fix,
the testcase added here just hangs, as we never handle proxied calls
to the main runtime thread (since we think we are never that thread).

Fixes #11914 (testcase in #12080)