Remove STATIC* JS vars and STATIC_BUMP setting (#12176)
These used to make sense when we had static allocations in JS. We'd use the
"bump" to track the total size, which was increased by the JS compiler, and at
runtime we'd have STATIC_BASE and STATICTOP. Now that we disallow static
allocations from the JS compiler, we don't need them.

Note that we do still track "staticBump" in the metadata from finalize. That is
the total size of static memory from lld, which never changes. We use it to
compute where the stack begins (which determines the rest of memory layout).
As a followup we may be able to use a link map or other mechanism to just
read that info instead of computing it in emscripten.py

See #11860