Update emsize test case
The contents of the test `.js` and `.wasm` were quite old
and had refenences to legacy names such as DYNAMICTOP_PTR.
This isn't really problem but it makes the output of `git grep`
a little cleaners.  Its also probably good to update this
from time to time so we know we are testing actual output.