Fix pthread_cleanup_push stub function
A typo was introduced to the dependencies of pthread_cleanup_push
in #1205.  However, we can do better by simply avoiding the dynCall
here.

Also, remove _pthread_cleanup_push and _pthread_cleanup_pop variants
that we don't use in emscripten and if we were to implement them
they would require a different signature.

Fixes: https://github.com/emscripten-core/emsdk/issues/614