Run wasm-emscripten-finalize only when needed, and add ERROR_ON_WASM_CHANGES_AFTER_LINK option (#12173)
emscripten.py now tracks whether we actually need to run wasm-emscripten-finalize.
Currently we don't need to in a specific subset of -O0 builds, in which we don't
need legalization (by using BigInt support) and we don't need invokes (which means
we don't use longjmp or C++ exceptions; this restriction will be lifted once
WebAssembly/binaryen#3081 is fixed).

This new option will show an error if any binaryen work
is done after wasm-ld. That is, it ensures that we do not run wasm-emscripten-finalize
or wasm-opt. That is useful to verify that the link is as fast as possible and also that
it does not rewrite DWARF, which is necessary to support split DWARF (binaryen can
rewrite normal DWARF, but it's simpler and better to support split DWARF by just
not doing any binaryen work after wasm-ld).

See WebAssembly/binaryen#3043