Do not run binaryen opts in -O1 (#12178)
Suggested in #12173 (review)

This appears to regress code size by around 5-10% on average for -O1. Given
that this mode is kind of a "compromise" build anyhow, with most opts in return
for fast iteration, it seems like a better tradeoff to lose that 5-10% in return for
much faster link times - almost 2x faster - and also avoiding DWARF rewriting.