Stop including malloc/free by default (#12213)
Before this we would always include malloc/free, and rely on metadce to remove them
when not needed. But this has downsides:

1. Slower compile times.
2. Larger builds when not using metadce.
3. It prevented us from writing a nice sbrk using __heap_base, as
if malloc is included first and removed by metadce that would still
leave some data in the memory initialization, increasing code size
without the workaround we had in sbrk to avoid it.

Much of the work here is to get deps_info.json to handle things properly and
include malloc/free when needed. I did this twice to double-check, but it is still
someone error-prone, as we need each code path that reached a malloc from
JS to be identified properly.

Some places use malloc/free in locations that are not part of the JS deps
system, and those now use makeMallocAbort. In those places we know
we should not reach them if malloc is not included, so we just have an
abort there. That would catch errors with deps_info.json.

In assertions builds, also help users by adding malloc/free in JS that show
a clear error if called, if malloc/free was not included but the user called
them anyhow.

With this PR we can remove the hack for sbrk, but I wanted to keep that
for a later PR as this is already not small.