Remove use of tableBase in relocateExports
We don't export `fp$` function from native code either with
llvm backend or with fastcomp as fast as I can tell.

The `fp$` functions are only ever imported *imported*.  As
far as I can tell this has always been true so I'm not sure
why this code was originally added in #6334.