Use getentropy from std::random_device, avoiding FS usage of /dev/urandom (#12240)
This does not change the random number mechanism, it just avoids going through the FS
in order to get random values. getRandomDevice is refactored out of the FS code without
change into a global helper. At the C level we now support getentropy() which is what
libc++ calls.

By avoiding the FS this is faster, in particular in pthreads builds all FS operations are
proxied, so this PR avoids that overhead.

As a benefit this makes us do the same as WASI, so it removes a difference.

Also make the test for this actually get a random number, and not just set up the device
(that used to open the FD, so the test wasn't useless before).

Fixes #10068.