Clean up pthread startup code (NFC) (#12308)
There were two places that pthread workers were initialized, one in a postSet,
and one at the end of postamble. This joins the two into a single initWorker
method, and moves code to there, which is simpler.

(Also add a missing initWorker for minimal runtime, so this is technically
not NFC for that runtime. Note that there may be more fixes needed there,
this just does the obvious thing for this PR to improve things.)

This will make fixing a race condition in a subsequent PR easier.