tests: Removal of @no_wasm_backend, part 1
This change removes the use of this decorator from the other
and sanity test suites.

Some tests simply worked, some make no sense with wasm backend and
a couple are tests we would like revive once day (related to ctor
evaluation and ctor elimiation).

There was one test for which I implemented the feature which was
to ingnore multiple copies of the same .so file on the link line.
This improves the fake shared library support a little and was
a feature of fastcomp.

See: #12335