Rewrite minifyGlobals JS optimizer pass (#12329)
This fixes a bug that only became apparent when wasm2js started to
emit more interesting JS code, which confused the pass. Historically
the pass was just used on asm.js globals, which are trivial, and it was
reused for wasm2js, but as wasm2js output changed we ran into the
issue. Specifically, the old code made some assumptions about
fastcomp output like that the first item shouldn't be minified, and that
we can ignore some things and not others; those hacks are now all
removed.

The new design is cleaner and more general, but still not 100% fully
general, see the corner case it doesn't handle in the comment. It may
be good to handle that to prevent future bugs, but it would add
significant complexity here. Adding just an assert might be better,
but that would still need the complexity to reason about scopes.

Unintentionally, this cleaner design can also do better at minifying,
improving wasm2js hello world's total size by over 10% (!)

This removes the old fastcomp-style handling in minifyGlobals, with the
testcase for that. The new pass just focuses on wasm2js's output
format.

Once this lands we can re-enable minification in binaryen.js CI, which
is disabled atm.