Add EM_COMPILER_WRAPPER as replacement for EMMAKEN_COMPILER
This allows a compiler wrapper to be installed around the
underlying clang, but without changing the location of the
llvm installation or having to know where it is.

This is in contrast to other option LLVM_ROOT which points
emscripten at a different llvm build (and also replaces not
just clang but all the other llvm tools).

This mimikes the corresponding CMAKE option
CMAKE_<LANG>_COMPILER_WRAPPER.

The reason we allow the inner compiler to be wrapped like this
is that distributing the whole of emscripten to a distcc farm
or gomacc farm would otherwise be required.

Fixes: #12340