Update code size tests for recent improvements (#12390)
WebAssembly/binaryen#3163 removed some code and this was enough to break one
of the rollers, where we were on the edge apparently. But most of the recent
wins have just accumulated without causing an error.

This updates to the new expectations including that PR, which shows the nice improvements.
It also increases the slop for now just so we can pass on CI before and after. Will remove
that after.