If async preparing wasm fails with MODULARIZE=1, reject the module ready promise
For example, if a global constructor performs an invalid memory access,
an exception will be thrown that bubbles up to receiveInstantiatedSource.
That causes a rejection of the promise returned from instantiateAsync,
and we should forward that to the module consumer via the module ready promise.
By handling that rejection we also avoid an unhandled-promise-rejection
error in browsers.

Fixes #12396