Disable browser.test_pthread_create with minimal runtime, which is flakey (#12383)
See #12368 - the flakiness in the test appears to be 100% due to the minimal runtime
part, so something specific is racy there. After spending a little time I saw nothing
obvious, but maybe I'm not familiar enough with minimal runtime to know where to
look, that is, where the differences between the runtimes would be regarding pthreads.

For now this disables that test on minimal runtime, and also makes the test stricter -
it's 100% deterministic so there's no need to retry once after a failure (which actually
made the flakiness here harder to notice).

Note that this does not remove all minimal runtime + pthreads tests, as there is still
test_minimal_runtime_hello_pthread (which is simpler, and shows no flakiness).