Remove all unused js optimizer passes (#12414)
This removes some TODOs from wasm2js where we thought about
running more of the old passes. I think that's not a good idea - they
will not run as our wasm2js output gets less and less like asm.js
anyhow. We'll need to write new acorn passes there, if we want to.

This removes safeHeap which we should be running (see #12410 (comment)),
but as it would be the only thing we use the js optimizer for from emcc.py,
and that isn't a recent regression, remove that too and I'll rewrite it for
acorn.

See https://github.com/emscripten-core/emscripten/issues/11860