Improve SUPPORT_LONGJMP=0 (#12394)
Don't run the backend pass if longjmp support is disabled. This allows setting the
flag at compile time to disable invoke generation for longjmps, when that is not
desired.

Also, before this change, the error would just be "undefined symbol: longjmp" which is
not that clear, and worse, if errors on undefined symbols were disabled, it would
just be a warning. This emits a clear error at link time if the flag is set to not support
longjmp but the code requires it.