Add clang-format to github actions
I would have prefered to use an actual pre-packaged github
action for this.  There seem to be very many of these but
none of them seems to have the most important feature here
which is the ability to run just on the diff and not on the
entire repo.

This probably shouldn't land until we figure out exactly
what parts of the tree we want to run this on.

I'm thinking probably `tests` but also any part of `system`
that are emscripten-local.

We also might want to do a quick first pass at formatting
these things before landing this.