Simplify windows and mac builds (#12445)
They each had a single test job. Merge the build+test jobs into a single job,
which is simpler (less code in the circleci config) and also faster (avoids
the overhead of starting a second job, a second install of choco packages,
and persisting and unpersisting data).

Also remove gen_struct_info hack (as the hack actually breaks things
after the simplification, not splitting it out).