Use a mutable global as __stack_pointer in dynamic linking
Without this change we were using a normal Number and then
wasm-emscripten-finalize would use that value to set an
internal mutable global.

This change means that users emscripten's dynamic linking
now depends on the mutable globals proposal.  However this
is already available in all major browsers:

  https://webassembly.org/roadmap/

This is small step towards #12461 which can be landing
in isolation without otherwise changing the ABI.

Once this lands we can remove internalizeStackPointerGloba
from wasm-emscripten-finalize.