Fix dlsym for 64-bit functions
This basically involved looking the `orig$` functions whenever we
want a native WebAssembly function.

Regarding test_dlfcn_i64: It seems that this test wasn't actually
testing functions that have i64 in thier signature but some kind of i64
handling from back in the asm.js days.  See  #2060.  This change
repurposes it to and makes it more useful.