Add `--oformat` flag to explictly control output format
Without this flag the output format is determined by the name of the
output file.

With this flag the name of the output file is ignored and the format is
determined solely by this flag.

This can be useful in various situation but the motivation for doing
it now that it provides a place to specify a new output format which
will be the raw linker output, which can then be fed into that new
post-link-only mode that we are planning.