Force use of Ninja generator in test_cmake_with_embind_cpp11_mode (#12613)
This change includes a bunch of cleanups but the main change
here is to force the use of the Ninja generator.  Without
this this test will try to VS or nmake on windows which fails
in various ways.

This change is needed since we stopped forcing the default
generator on windows in #12595 landed.