Read static data size from dylink section
This allows us to stop depending wasm-emscripten-finalize to
write staticBump which in turn will allow us to stop exporting
`__data_end`, once we update binaryen.

As a side effect of this the static bump information is no
longer available in the normal linking mode which means we
can't include dynamic_base in EMSCRIPTEN_METADATA anymore.  One
more nail in the coffin for EMSCRIPTEN_METADATA. See #12231