Allow dlsym() to looking JavaScript symbols in RTLD_DEFAULT mode (#12609)
In this mode dlsym is supposed to be able to looking all symbols
in default library search path which should include JS symbols.

Unfortunately if the symbol being looked up is not already in the
table then its signature needs to be known at this point.  To
handle this case we keep the signatures around at runtime by
attaching them to the JS library functions.  This a small extra
overhead that only applies to dynamic linking.

This change is needed for and split out from #12461