Basic support --post-link mode running on output of wasi-sdk
This is an initial PR that only supports the most basic of
wasi-sdk-produced programs.  Most ones that just use stdout and argv.

Adds a simple test which uses the official wasi-sdk sysroot
to build the wasm binary and then uses emscripten to post-process
this to turn this into a runnable JS file.

See #12341