Error on bad JS library elements
For libary config element such as `foo__sig` error out if
the corresponding library member is not found (in this
case if `foo` is not in the library.

This found `memcpy__inline` which can be removed since
memcpy is no longer part of the JS library.

Also remove `__inline` completely since this was the last
use of it, and its asmjs specific.