Make the DOM codes reproducible
Whilst working on the Reproducible Builds effort [0] I noticed that emscripten
could not be built reproducibly.

This is because it generates a switch/case statement using Python's
random.randint/random.uniform functionality.

This patch seeds this random component using the SOURCE_DATE_EPOCH environment
variable [1] so that, iff this variable is present, it will deterministically
generate the same output.

I originally filed this bug in Debian as #973595 [2].

 [0] https://reproducible-builds.org/
 [1] https://reproducible-builds.org/specs/source-date-epoch/
 [2] https://bugs.debian.org/973595