Split out post-link code in emcc and make it independently runnable (#12640)
This change introduces a new output format (--oformat) called
`bare`.  In `bare` mode emscripten will stop after running wasm-ld
and output just the bare wasm module.

It also introduces a new command line argument `--post-link` which
take as input a raw wasm module and run all the post link phases
on it (e.g. wasm-emscripten-finalize, jsifier, etc).

The precise command line UI for this feature is still in the
design phase so I've added a warning so that any users of this
command line UI know that it is experiment.