Fix parsing of JS library pre-processor directives
Because we were only checking for `#if` at the start of the
line we were accidenatlly also accepting `#ifdef` and also
`#ifanything`.

For backwards compatibility continue to accept `#ifdef` in
case people are using it in the wild but issue warning.

As a side effect of this refactor we not allow support
preprocessor directives that are not in column 0.