Avoid temp files in compile-only mode
In compile-only mode we can write all object files directly to their
final locations completely avoid the need for temp files or for
renaming. Also avoid the `.d` file modification that we had in there.

This is a really nice simplification that has become possible as we have
made the emscripten code more precise about exactly which mode it is
running in.

Fixes: #12709