Limit line length of JS ouptut in test code (#12823)
When minified JS tests fail the resulting error messages from node
normally contain the entire program (since its all on the same line).
This didn't used to be the case with older versions of node IIRC.

This change used the exiting `limit_size` function when displaying JS
output and also limits the size of any give line to avoid flodding the
terminal with a million chars on a single line.

If folks find this annoying, or were relying on this output when running
tests we add options.  For example we could limit this behaviour to only
apply run running in a terminal.