Remove some dead pthread code
I'm not sure where MAIN_THREAD_ID is left over from but `1` is never
used as a pthread ID.   We have a special pthread ID that is registered
as the main thread using emscripten_register_main_browser_thread_id but
that is always a pointer which is the resolve of malloc, and never `1`.

This also means I guess we don't have tests for killing or canceling
main?