Use musl's pthhread_self.c. NFC.
This change make emscripten more like other platorms in that
`__pthread_self` is implementd in the header file with inline asm and
the public `pthread_self` function comes from pthread_self.c

Previously we had it kind of backwards where the low level
`__pthread_self` was calling into the higher level `pthread_self`.

This is also the first use of inline asm we have in emscripten
and its good to know that it works.