Remove PThread.mainThreadFutex from JavaScript. NFC (#12887)
This variable is initialized from `_main_thread_futex` which is a data
address exported from native code.  Rather than copy it we can simply
use that variable directly.

I think it made sense to have a separate JS variable in the past when
this address was allocated at runtime, but since #12318 landed it
is no longer needed.

Also rename the exported variable to avoid conflict with user symbols.