Ignore part of test_dwarf in unoptimized builds (#12924)
LLVM now emits larger code in unoptimized builds, with some extra local
assignments, instead of the simple pattern the test expected, which is to
set the local to the constant value and then do the call. Instead, it sets
the constants earlier on and then does some copying of those locals. I
can't see a simple way to parse that pattern in a useful way, so skip part
of the test in unoptimized builds. (If we think it's useful to test this kind of
thing, we'd need a different way entirely, probably, that isn't sensitive to
LLVM changes like this.)

Fixes #12903